Properly append LabelSelector
The value of StorageCluster.Spec.LabelSelector was not being properly appended
to the generated Placements of the OCS components.Specifically, adding a new
NodeSelectorTerm results in a selector term thatis ORed with any other terms,
whereas adding a newNodeSelectorRequirement to an existing NodeSelectorTerm
would result in an AND of all requirements in that term. This commit makes it
so the LabelSelector field is ANDed with any other existing node selectors in
the Placement.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>